requirements.txt
Django
djangorestframework
dj-rest-auth
django-allauth
pillow
django-cors-headers
email_validator
drf_writable_nested
